SI-7747 Added .INSTANCE accessor to classBasedWrappers in REPL.
The changes to wrappers are explained as follows.

$ scala -Yrepl-class-based

```scala

scala> val a = 10
a: Int = 10

scala> case class A(a: Int)
defined class A

scala> A(a) // show
class $read extends Serializable {
// ... some uninteresting stuff elided.
   class $iw extends Serializable {
// ...
    val $VAL1 = $line5.$read.INSTANCE; // <-- First points of interest.
    import $VAL1.$iw.$iw.a;           //  <--
    import $line6.$read.INSTANCE.$iw.$iw.A; // <- Second point of interest.
    import $line6.$read.INSTANCE.$iw.$iw.A;
    class $iw extends Serializable {
// ...
      val res0 = A(a)
    };
    val $iw = new $iw.<init>
  };
  val $iw = new $iw.<init>
}
object $read extends scala.AnyRef {
// ...
  val INSTANCE = new $read.<init>
}
res0: A = A(10)

scala>

```

(a) First point of interest.

(1) Originally -Yrepl-class-based would import the same thing as follows:
```scala
import $line6.$read.$iw.$iw.a;
```
(2) This patch changes it to following:
```scala
val $VAL1 = $line5.$read.INSTANCE; // <-- First points of interest.
import $VAL1.$iw.$iw.a;
```
From my understanding, In (1) we directly import a value `a`, when this code
is shipped to a remote executor leads to initializing all the classes wrapping
`a`, not just that `a` itself had lots of imports around it
( because we do not finesse on imports to allow implicits. ) when it was wrapped
. Then this leads to bubbling up
( in a chain initializing the classes involved in those imports.), way up to
where we initialized the first thing in the repl. All of these classes are
reinitialized on the remote repl. This is undesirable in most cases.

In (2) we break the import between a val pointing to an already created
instance and then importing it. This leads to importing the deserialized
instance of the accessor `.INSTANCE` and does not lead to bubbling up
phenomenon mentioned above.

 (b) Second point of interest.

It is as it was originally, this lead to adding a special case for class
declaration. In this patch it was changed to finesse on what we import,
since when we declare a class - implicit imports are not needed.
This fine list of import avoids referring to unnecessary things class
declaration needed. And the bubbling up phenomenon is avoided and not eliminated.
 If at all it happens, it does so in expected way.
This was not changed to above style(mentioned in (a)) to prevent isomorphism
that would result from importing from different val ids
(eg : `$VAL1` and `$VAL2` referring to same type would be treated as different).